# Write to rotating ContentReWriter in transformer chain