--- a/transform/absurl.go
+++ b/transform/absurl.go
@@ -17,8 +17,8 @@ func initAbsURLReplacer(baseURL string) {
 func AbsURL(absURL string) (trs []link, err error) {
 	initAbsURLReplacer(absURL)
 
-	trs = append(trs, func(content []byte) []byte {
-		return ar.replaceInHTML(content)
+	trs = append(trs, func(rw ContentReWriter) {
+		ar.replaceInHTML(rw)
 	})
 	return
 }
@@ -26,8 +26,8 @@ func AbsURL(absURL string) (trs []link, err error) {
 func AbsURLInXML(absURL string) (trs []link, err error) {
 	initAbsURLReplacer(absURL)
 
-	trs = append(trs, func(content []byte) []byte {
-		return ar.replaceInXML(content)
+	trs = append(trs, func(rw ContentReWriter) {
+		ar.replaceInXML(rw)
 	})
 	return
 }