--- a/transform/chain.go
+++ b/transform/chain.go
@@ -1,12 +1,12 @@
 package transform
 
 import (
-	"io"
-
+	"bytes"
 	bp "github.com/spf13/hugo/bufferpool"
+	"io"
 )
 
-type trans func([]byte) []byte
+type trans func(rw ContentReWriter)
 
 type link trans
 
@@ -20,17 +20,62 @@ func NewEmptyTransforms() []link {
 	return make([]link, 0, 20)
 }
 
-func (c *chain) Apply(w io.Writer, r io.Reader) (err error) {
-	buffer := bp.GetBuffer()
-	defer bp.PutBuffer(buffer)
+// ContentReWriter is an interface that enables rotation
+// of pooled buffers in the transformer chain.
+type ContentReWriter interface {
+	Content() []byte
+	io.Writer
+}
 
-	buffer.ReadFrom(r)
-	b := buffer.Bytes()
-	for _, tr := range *c {
-		b = tr(b)
+// Implements ContentReWriter
+// Content is read from the from-buffer,
+// and rewritten to to the to-buffer.
+type fromToBuffer struct {
+	from *bytes.Buffer
+	to   *bytes.Buffer
+}
+
+func (ft fromToBuffer) Write(p []byte) (n int, err error) {
+	return ft.to.Write(p)
+}
+
+func (ft fromToBuffer) Content() []byte {
+	return ft.from.Bytes()
+}
+
+func (c *chain) Apply(w io.Writer, r io.Reader) error {
+
+	b1 := bp.GetBuffer()
+	defer bp.PutBuffer(b1)
+
+	b1.ReadFrom(r)
+
+	if len(*c) == 0 {
+		b1.WriteTo(w)
+		return nil
 	}
-	buffer.Reset()
-	buffer.Write(b)
-	buffer.WriteTo(w)
-	return
+
+	b2 := bp.GetBuffer()
+	defer bp.PutBuffer(b2)
+
+	fb := &fromToBuffer{from: b1, to: b2}
+
+	for i, tr := range *c {
+		if i > 0 {
+			if fb.from == b1 {
+				fb.from = b2
+				fb.to = b1
+				fb.to.Reset()
+			} else {
+				fb.from = b1
+				fb.to = b2
+				fb.to.Reset()
+			}
+		}
+
+		tr(fb)
+	}
+
+	fb.to.WriteTo(w)
+	return nil
 }